waypoint -> checkpoint
A waypoint is "an endpoint of the leg of a course, esp. one whose coordinates
have been generated by a computer". i.e. it's what we are generating.

A checkpoint is "a location whose exact position can be verified visually or
electronically, used by pilots to aid navigation"

So the stuff in {apt,nav,fix}.dat.gz are checkpoints, and the endpoints of our
course legs are waypoints.

darcs-hash:20070521001041-28bda-95b70e44f3b18afd0094dabd0953458ee373270d.gz